Create MatchStickGame.c
This is a game in which there are 21 matchsticks and user and computer will play it turn by turn. Both of them can pick a certain number of matchsticks at once i.e. from 1 to 4. The who will be forced to pick up the last matchstick will lose the game.
Note: This program is designed specially so that computer will win every single time.